# backend/src/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Coxa Fan OS API
  version: "0.2.0"
  description: |
    REST API for the Coxa Fan Operating System.
    Covers authentication, retail POS, ticketing, fan CDP, loyalty, and personalization.
    ## Authentication
    Obtain a JWT from POST /auth/login then pass Authorization: Bearer token.
    ## Tenant Context
    Every request needs x-tenant-id header (club identifier, e.g. coxa-club-001).
    ## Rate Limits
    Auth: 30/15min | Ticket ops: 60/min | Default: 500/min.
servers:
  - url: /api/v1
    description: Current API version
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantHeader:
      type: apiKey
      in: header
      name: x-tenant-id
...
  schemas:
...
    SaleLineQr:
      type: object
      properties:
        qrToken:
          type: string
        unitIndex:
          type: integer
        status:
          type: string
          enum: [issued, redeemed, voided]
        redeemedAt:
          type: string
          format: date-time
    MatchEvent:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        matchDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, on_sale, sold_out, cancelled, completed]
    Ticket:
      type: object
      properties:
        _id:
          type: string
        matchEventId:
          type: string
        fanProfileId:
          type: string
        qrToken:
          type: string
        seat:
          type: string
        status:
          type: string
          enum: [issued, used, cancelled]
...
    CdpEvent:
      type: object
      properties:
        _id:
          type: string
        eventName:
          type: string
        source:
          type: string
        fanProfileId:
          type: string
        idempotencyKey:
          type: string
        eventTimestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [accepted, processed, dlq]
        payload:
          type: object
...
security:
  - bearerAuth: []
  - tenantHeader: []
paths:
...
  /retail/sales/{saleId}/qr-codes:
    get:
      tags: [Retail]
      summary: Get sale QR codes
      description: Returns per-unit QR tokens for a completed sale, grouped by sale line.
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR codes grouped by sale line
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      saleId:
                        type: string
                      saleNumber:
                        type: string
                      lines:
                        type: array
                        items:
                          type: object
                          properties:
                            saleLineIndex:
                              type: integer
                            productName:
                              type: string
                            tokens:
                              type: array
                              items:
                                ref: "#/components/schemas/SaleLineQr"
                      totalTokens:
                        type: integer
  /retail/sale-qr/redeem:
    post:
      tags: [Retail]
      summary: Redeem sale QR
      description: Marks a single per-unit QR token as redeemed at the point of delivery.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qrToken]
              properties:
                qrToken:
                  type: string
      responses:
        "200":
          description: QR redeemed
        "409":
          description: Already redeemed or voided
        "404":
          description: QR token not found

  # ── Ticketing ──────────────────────────────────────
  /ticketing/events:
    get:
      tags: [Ticketing]
      summary: List match events
      parameters:
        - name: upcoming
          in: query
          schema:
            type: boolean
          description: If true, returns only future events
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, on_sale, sold_out, cancelled, completed]
      responses:
        "200":
          description: List of match events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      ref: "#/components/schemas/MatchEvent"
    post:
      tags: [Ticketing]
      summary: Create match event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, matchDate]
              properties:
                name:
                  type: string
                homeTeam:
                  type: string
                awayTeam:
                  type: string
                matchDate:
                  type: string
                  format: date-time
                venueId:
                  type: string
      responses:
        "201":
          description: Match event created
  /ticketing/events/{id}:
    get:
      tags: [Ticketing]
      summary: Get match event detail
      description: Returns event with ticket products and check-in windows.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail with products and windows
        "404":
          description: Event not found
  /ticketing/events/{id}/status:
    patch:
      tags: [Ticketing]
      summary: Update event status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [scheduled, on_sale, sold_out, cancelled, completed]
      responses:
        "200":
          description: Status updated
  /ticketing/events/{id}/ticket-products:
    get:
      tags: [Ticketing]
      summary: List ticket products for event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket products (tiers/sections)
    post:
      tags: [Ticketing]
      summary: Create ticket product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, priceCents, totalQty]
              properties:
                name:
                  type: string
                  example: Section A - General
                priceCents:
                  type: integer
                totalQty:
                  type: integer
      responses:
        "201":
          description: Ticket product created
  /ticketing/tickets:
    get:
      tags: [Ticketing]
      summary: List tickets
      parameters:
        - name: fanProfileId
          in: query
          schema:
            type: string
        - name: matchEventId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [issued, used, cancelled]
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      ref: "#/components/schemas/Ticket"
...
  /ticketing/entitlements/validate:
    get:
      tags: [Ticketing]
      summary: Validate QR (GET)
      description: Validate an entry QR token at the gate. Query param version.
      parameters:
        - name: qrToken
          in: query
          required: true
          schema:
            type: string
        - name: matchEventId
          in: query
          schema:
            type: string
        - name: markUsed
          in: query
          schema:
            type: boolean
        - name: gateId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Validation result
    post:
      tags: [Ticketing]
      summary: Validate QR (POST)
      description: Validate an entry QR token at the gate. Body version — preferred for POS/gate hardware.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qrToken]
              properties:
                qrToken:
                  type: string
                matchEventId:
                  type: string
                markUsed:
                  type: boolean
                  default: false
                gateId:
                  type: string
                deviceId:
                  type: string
      responses:
        "200":
          description: Validation result
        "404":
          description: QR token not found
        "409":
          description: Already used
...
  /cdp/events:
    get:
      tags: [CDP]
      summary: List CDP events
      parameters:
        - name: eventName
          in: query
          schema:
            type: string
            example: sale.completed
        - name: fanProfileId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [accepted, processed, dlq]
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of CDP events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      ref: "#/components/schemas/CdpEvent"
    post:
      tags: [CDP]
      summary: Ingest CDP event
      description: |
        Publishes a domain event into the CDP pipeline.
        Supported event names: sale.completed, fan.registered, ticket.purchased,
        ticket.used, member.checked_in, no_show.recorded, loyalty.points.earned.
        Idempotent — same idempotencyKey returns the existing event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventName, source, idempotencyKey]
              properties:
                eventName:
                  type: string
                  example: sale.completed
                source:
                  type: string
                  example: retail_pos
                idempotencyKey:
                  type: string
                fanProfileId:
                  type: string
                fanEmail:
                  type: string
                eventTimestamp:
                  type: string
                  format: date-time
                payload:
                  type: object
      responses:
        "201":
          description: Event ingested
        "200":
          description: Duplicate — existing event returned
...